+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -39,7 +39,7 @@ public class UserService {
         return userDao.update(user);
     }
 
-    public User getUserById(int userId) {
+    public User getUserById(Long userId) {
         isUserExists(userId);
         return userDao.getUserById(userId);
     }
@@ -48,21 +48,21 @@ public class UserService {
         return userDao.getAllUsers();
     }
 
-    public void addFriend(int userId, int friendId) {
+    public void addFriend(Long userId, Long friendId) {
         isUserExists(userId);
         isUserExists(friendId);
         feedDao.create(new Feed(userId, EventType.FRIEND, Operation.ADD, friendId));
         friendDao.saveFriend(userId, friendId);
     }
 
-    public void deleteFriend(int userId, int friendId) {
+    public void deleteFriend(Long userId, Long friendId) {
         isUserExists(userId);
         isUserExists(friendId);
         feedDao.create(new Feed(userId, EventType.FRIEND, Operation.REMOVE, friendId));
         friendDao.deleteFriend(userId, friendId);
     }
 
-    public List<User> getFriendsByUserId(int userId) {
+    public List<User> getFriendsByUserId(Long userId) {
         isUserExists(userId);
         return friendDao.getFriendsByUserId(userId)
                 .stream()
@@ -70,7 +70,7 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
-    public List<User> getCommonFriends(int userId, int otherUserId) {
+    public List<User> getCommonFriends(Long userId, Long otherUserId) {
         List<User> friends = getFriendsByUserId(userId);
         List<User> otherFriends = getFriendsByUserId(otherUserId);
         return friends.stream()
@@ -78,25 +78,25 @@ public class UserService {
                 .collect(Collectors.toList());
     }
 
-    public List<Feed> getAllFeedsByUserId(int id) {
+    public List<Feed> getAllFeedsByUserId(Long id) {
         return feedDao.getAllFeedsByUserId(id);
     }
 
-    public List<Film> getFilmRecommendations(Integer userId) {
+    public List<Film> getFilmRecommendations(Long userId) {
         if (!userDao.containsInStorage(userId)) {
             throw new NotFoundException("User with id = " + userId + " not found");
         }
         return filmDao.getFilmRecommendations(userId);
     }
 
-    public void deleteUserById(int userId) {
+    public void deleteUserById(Long userId) {
         if (!userDao.containsInStorage(userId)) {
             throw new NotFoundException("User with id = " + userId + " not found");
         }
         userDao.deleteUserById(userId);
     }
 
-    private void isUserExists(Integer userId) {
+    private void isUserExists(Long userId) {
         if (!userDao.containsInStorage(userId)) {
             throw new NotFoundException(String.format("User with id = %d not found", userId));
         }
